Improvement to the Test Framework to test block processing The current test framework does not allow new test cases to add blocks to the block-chain. This was pointed out by Mike Hearn on the development list in a thread with subject "[Bitcoin-development] Question on creating test cases for block.CheckBlock()"
After debugging the Bitcoin core, we found that this is because of three reasons:
1. The miner_tests.cpp breaks the block-chain and assumes it's empty on start. So other test cases cannot be run neither before not after.
2. Creating blocks on the fly requires computing proof-of-work. Even if the starting difficulty is low (1) it is too much for a dynamic testcase to solve.
3. Test cases cannot start with a fresh new block-chain because other test cases may have previously added blocks.

We belive that testing the block acceptance rules is crucial for the safety of the application and so we wrote this patch.

This patch improves the testing framework by solving 1,2,3 with:
- Providing a method to reset the block-chain to the starting state (testingSetupManager.SetupGenesisBlockChain())
- Allowing to skip the proof-of-work testing (supressCheckBlockWork = true)

We've also found the exact procedure that can be used dynamically to destroy a block-chain correctly, which was undocumented, and some parts missing in the code.
This could be of great help to re-create completely the block-chain in case a severe damage has been detected, without restarting the appication.

As a bonus, 7 test cases have been added:
- ToCheckBlockUpgradeMajority (untested before)
- EnforceBlockUpgradeMajority (untested before)
- RejectBlockOutdatedMajority (untested before)
- "bad-cb-height"
- "bad-version"
- "time-too-old"
- "bad-txns-nonfinal"

Last, we added a way to leave the block-chain unaltered after the test suite is over to debug the test cases themselves (testingSetupManager.keepTestEvidence = true)

S.D.L. & T.H.